Fixed some wrong Unicode formats and added more
Some of the Unicode formats were wrong on the last commit, so in this one that issue has been fixed.

Also, some more questions have been corrected, but some will remain uncorrected due to the incapability of fully understanding what the question was supposed to day.

Note: this was already done in stable before the DevTools was implemented.